+++ src/hooks/useSelectedCartIds.ts
@@ -2,11 +2,27 @@ import { useState } from "react";
 import { CartProduct } from "../type/cart";
 
 const useSelectedCartIds = (cartItems: CartProduct[]) => {
+  const LOCAL_STORAGE_KEY = "selectedCartIds";
+
+  const getInitialSelectedIds = (): number[] => {
+    const stored = localStorage.getItem(LOCAL_STORAGE_KEY);
+    if (stored) {
+      return JSON.parse(stored);
+    }
+
+    return cartItems.map((item) => item.id);
+  };
+
   const [selectedCartIds, setSelectedCartIds] = useState<number[]>(
-    cartItems.map((item) => item.id)
+    getInitialSelectedIds
   );
 
-  return { selectedCartIds, setSelectedCartIds };
+  const handleSelectedCartIds = (newSelectedCartIds: number[]) => {
+    setSelectedCartIds(newSelectedCartIds);
+    localStorage.setItem(LOCAL_STORAGE_KEY, JSON.stringify(newSelectedCartIds));
+  };
+
+  return { selectedCartIds, handleSelectedCartIds };
 };
 
 export default useSelectedCartIds;
